requirements.txt
masumi>=0.1.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
openai>=1.0.0
tavily-python>=0.3.0
pydantic>=2.0.0
uvicorn>=0.23.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0
